requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.4.0
Flask-Cors==4.0.0
python-dotenv==1.0.0

# Database
SQLAlchemy>=2.0.16
# psycopg2-binary==2.9.0  # Commented out for Windows + MySQL workflow; add only if PostgreSQL is required
mysql-connector-python==8.0.33

# Deep Learning & AI (optional; install separately when your environment is validated)
# tensorflow==2.21.0
# torch==2.11.0
# torchvision==0.17.2
# scikit-learn==1.3.2
# numpy==1.24.0
# pandas==2.0.0
# Pillow==9.5.0

# NLP & Chatbot (Optional; requires Rust for tokenizers and heavy packages)
# transformers==4.30.0
# huggingface-hub==0.15.0
# nltk==3.8.0

# IoT & Sensors
paho-mqtt==1.6.1
requests==2.31.0

# Notifications & Alerts
twilio==8.2.0
email-validator==2.0.0

# Data Processing & Reports
openpyxl==3.1.5
reportlab==4.4.10
PyPDF2==3.0.0

# Geospatial
geopy==2.3.0
folium==0.14.0

# Utils
python-dateutil==2.8.0
pytz==2023.3
gunicorn==20.1.0
